Move attribute parsing out of data module
